+++ contracts/bridge/contracts/test/MockDarwiniaLightClient.sol
@@ -26,8 +26,8 @@ contract MockDarwiniaLightClient is LaneDataScheme {
     function verify_messages_proof(
         bytes32 outboundLaneDataHash,
         bytes32 inboundLaneDataHash,
-        uint256 chain_pos,
-        uint256 lane_pos,
+        uint32 chain_pos,
+        uint32 lane_pos,
         bytes calldata proof
     ) external view returns (bool) {
         bytes32 lane_hash = hash(LaneData(outboundLaneDataHash, inboundLaneDataHash));
@@ -37,8 +37,8 @@ contract MockDarwiniaLightClient is LaneDataScheme {
     function verify_messages_delivery_proof(
         bytes32 outboundLaneDataHash,
         bytes32 inboundLaneDataHash,
-        uint256 chain_pos,
-        uint256 lane_pos,
+        uint32 chain_pos,
+        uint32 lane_pos,
         bytes calldata proof
     ) external view returns (bool) {
         bytes32 lane_hash = hash(LaneData(outboundLaneDataHash, inboundLaneDataHash));
